Added binary compatibility validator
Resolves #305
