+++ build.gradle.kts
@@ -3,6 +3,8 @@ import org.jetbrains.kotlin.gradle.tasks.KotlinCompile
 
 plugins {
     kotlin("jvm")
+    id("org.jetbrains.kotlinx.binary-compatibility-validator") version "0.13.0"
+
 
     `kotlin-dsl`
